refactor restructure project directories for domain separation
