+++ src/main/java/com/example/demo/message/dto/MessageDto.java
@@ -1,4 +1,4 @@
-package com.example.demo.dto;
+package com.example.demo.message.dto;
 
 import lombok.Getter;
 import lombok.Setter;
