+++ src/main/java/com/example/demo/message/repository/MessageRepository.java
@@ -1,6 +1,6 @@
-package com.example.demo.repository;
+package com.example.demo.message.repository;
 
-import com.example.demo.domain.Message;
+import com.example.demo.message.domain.Message;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MessageRepository extends JpaRepository<Message, Long> {
