+++ src/main/java/com/example/demo/scene/dto/SceneDto.java
@@ -1,4 +1,4 @@
-package com.example.demo.dto;
+package com.example.demo.scene.dto;
 
 import lombok.Getter;
 import lombok.Setter;
