+++ src/main/java/com/example/demo/scene/repository/SceneRepository.java
@@ -1,6 +1,6 @@
-package com.example.demo.repository;
+package com.example.demo.scene.repository;
 
-import com.example.demo.domain.Scene;
+import com.example.demo.scene.domain.Scene;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SceneRepository extends JpaRepository<Scene, Long> {
