+++ src/main/java/com/example/demo/user/dto/UserDto.java
@@ -1,4 +1,4 @@
-package com.example.demo.dto;
+package com.example.demo.user.dto;
 
 import lombok.Getter;
 import lombok.Setter;
