+++ src/main/java/com/example/demo/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.example.demo.repository;
+package com.example.demo.user.repository;
 
-import com.example.demo.domain.User;
+import com.example.demo.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, String> {
