+++ src/main/java/com/example/demo/user/service/UserService.java
@@ -1,7 +1,7 @@
-package com.example.demo.service;
+package com.example.demo.user.service;
 
-import com.example.demo.domain.User;
-import com.example.demo.repository.UserRepository;
+import com.example.demo.user.domain.User;
+import com.example.demo.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
